otel: experimenting with receivertest
Use the receivertest.CheckConsumeContract to test filebeatreceiver. The api
does not quite work right now because it wants to look for a unique log id
in the attributes, which we don't use for otelconsumer. I'm working on
extending the api to support a callback function that returns the unique id
from a user defined function, this would allow us to use a body field
(`at metadata._id`).