otel: experimenting with receivertest
Use the receivertest.CheckConsumeContract to test filebeatreceiver. Requires
some piping to forward the "test_id" attribute required by receivertest.

There is an issue I don't fully understand, when receivertest returns a
non-permanent error it calls `batch.Retry()`, but after that the test goes
on until a time out, otelconsumer is not called as part of the retry process.